GetAttrPointer Check



# message auto-generated for no-merge-commit merge:
!2593 merge master into master

GetAttrPointer Check

Created-by: lcttie
Commit-by: lcttie
Merged-by: cann-robot
Description: ## 描述
基于邮件 "【风险排查整改】请CMC各组接口人对算子获取整型属性潜在风险排查整改" 对 MC2 A3/5 barrier, combine, dispatch 中涉及 GetAttrPointer 的位置进行了风险排查

## 关联的Issue
无

## 测试
黄区编译、静态检查、二级冒烟、蓝区冒烟、本地验证

## 文档更新
无

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [x] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: cann/ops-transformer!2593

+++ mc2/moe_distribute_combine_v2/op_host/op_tiling/moe_distribute_combine_v2_tiling.cpp
@@ -1374,7 +1374,7 @@ static void UbUsedCal(const uint64_t ubSize, const gert::TilingContext* context,
     bool enableSpecialExpert = (constExpertNum + zeroExpertNum + copyExpertNum > 0U);
     auto expandXDesc = context->GetInputDesc(config.expandXIndex);
     auto attrs = context->GetAttrs();
-    auto commQuantModePtr = attrs->GetAttrPointer<int>((config.attrCommQuantModeIndex));
+    auto commQuantModePtr = attrs->GetAttrPointer<int64_t>((config.attrCommQuantModeIndex));
     uint32_t maxSizeTokenBuf = (axisH * sizeof(expandXDesc->GetDataType()) + UB_ALIGN - 1) / UB_ALIGN * UB_ALIGN;
     uint32_t hExpandXTypeSize = axisH * sizeof(expandXDesc->GetDataType());
     uint32_t activeMaskAlignSize = axisBS * ((axisK * sizeof(bool) + UB_ALIGN - 1) / UB_ALIGN * UB_ALIGN);
